Create Company Relocation
This report analyzes and compares key data points for four major cities in Africa - Kampala (Uganda), Nairobi (Kenya), Lagos (Nigeria) and Bujumbura (Burundi) - to help FreeBorn Inc. determine the best city to relocate their headquarters. The analysis includes two main components which are the emissions data for sulfur dioxide, particulate matter, nitrogen dioxide, carbon monoxide and ozone. Data for the emissions was provided thus was only cleaned and filtered, analyzed and visualized. This provides insight into the air quality and environmental condition of each city. The other data is cost of living data which was collected from Numbeo & Expatistan covering rental costs in the cities, utilities, health care. Education and groceries. The data was cleaned, filtered and visualized to compare the affordability of each location.
The combined data analysis shows that Lagos, Nigeria is the most favorable location based on the data. Nigeria has moderate emissions levels and a relatively low cost of living compared to other cities examined. Bujumbura and Nairobi also show a promise with lower emissions than the megacity of Kampala. The report provides FreeBorn Inc. with a data driven assessment to help inform their decision on the optimal city in Africa to establish their new headquarters. The detailed visualizations and comparisons give a clear picture of the environmental conditions and costs associated with their location.